adjust configuration basing on subjective experience
It seems that testing datasets differ dramaticaly from real usage.
Generated configuration is visibly awful. It is way to spacious
so you should repeat and repeat the same correction dozens of
times before the algorithm adjusts.
Besides, the new configuration performs only a little bit worse
on tests than the 'optimal' one. And it is much better subjectively.
Science sucks!